chore: add image and collection controllers

+++ src/server/api/controllers/collection-controller.ts
@@ -3,9 +3,11 @@ import { createAuthorizedApiInstance } from "@/server/api/api-instance";
 import {
     createResponseFromResult,
     createResponseFromVoidResult,
+    UnauthorizedResponse,
 } from "@/server/api/controller-model";
 import { permissionMiddleware } from "@/server/core/middleware/permission-middleware";
 import { CollectionService } from "@/server/core/services/collection-service";
+import { ServerPermissionService } from "@/server/core/services/server-permission-service";
 import { createContextLogger } from "@/utils/context-logger";
 import { sValidator } from "@hono/standard-validator";
 import { z } from "zod";
@@ -39,15 +41,32 @@ const InviteSchema = z.object({
 export const collectionController = createAuthorizedApiInstance()
     .get("/:id", permissionMiddleware(Resource.COLLECTION, Action.READ), async c => {
         const id = c.req.param("id");
+        const user = c.get("user");
         logger.debug({ id }, "Getting collection request");
+
+        const hasPermission = await ServerPermissionService.hasCollectionAccess({
+            user,
+            collectionId: id,
+            action: Action.READ,
+        });
+
+        if (!hasPermission) {
+            logger.info({ id }, "User does not have permission for this action in collections");
+            return UnauthorizedResponse(c, {
+                message: "User does not have permission for this action",
+            });
+        }
 
         const result = await CollectionService.getCollection(id);
         return createResponseFromResult(c, result);
     })
     .get("/", permissionMiddleware(Resource.COLLECTION, Action.READ), async c => {
         logger.debug("Getting collections request");
+        const user = c.get("user");
 
-        const result = await CollectionService.getCollections();
+        const result = await CollectionService.getCreatedCollectionsByOrganization(
+            user.organizationId
+        );
         return createResponseFromResult(c, result);
     })
     .post(
@@ -71,6 +90,18 @@ export const collectionController = createAuthorizedApiInstance()
             const data = c.req.valid("json");
             logger.debug({ id, data }, "Updating collection request");
 
+            const hasPermission = await ServerPermissionService.isCollectionOwner({
+                user: c.get("user"),
+                collectionId: id,
+            });
+
+            if (!hasPermission) {
+                logger.info({ id }, "User does not have permission to update this collection");
+                return UnauthorizedResponse(c, {
+                    message: "User does not have permission for this action",
+                });
+            }
+
             const result = await CollectionService.updateCollection({ id, ...data });
             return createResponseFromResult(c, result);
         }
@@ -83,6 +114,18 @@ export const collectionController = createAuthorizedApiInstance()
             const id = c.req.param("id");
             const { imageIds } = c.req.valid("json");
             logger.debug({ id, imageIds }, "Adding images to collection request");
+
+            const hasPermission = await ServerPermissionService.isCollectionOwner({
+                user: c.get("user"),
+                collectionId: id,
+            });
+
+            if (!hasPermission) {
+                logger.info({ id }, "User does not have permission to update this collection");
+                return UnauthorizedResponse(c, {
+                    message: "User does not have permission for this action",
+                });
+            }
 
             const result = await CollectionService.addImages({ collectionId: id, imageIds });
             return createResponseFromResult(c, result);
@@ -97,6 +140,18 @@ export const collectionController = createAuthorizedApiInstance()
             const { imageIds } = c.req.valid("json");
             logger.debug({ id, imageIds }, "Removing images from collection request");
 
+            const hasPermission = await ServerPermissionService.isCollectionOwner({
+                user: c.get("user"),
+                collectionId: id,
+            });
+
+            if (!hasPermission) {
+                logger.info({ id }, "User does not have permission to update this collection");
+                return UnauthorizedResponse(c, {
+                    message: "User does not have permission for this action",
+                });
+            }
+
             const result = await CollectionService.removeImages({ collectionId: id, imageIds });
             return createResponseFromResult(c, result);
         }
@@ -108,15 +163,43 @@ export const collectionController = createAuthorizedApiInstance()
         async c => {
             const id = c.req.param("id");
             const { organizationId } = c.req.valid("json");
-            logger.debug({ id, organizationId }, "Inviting user to collection request");
+            logger.debug({ id, organizationId }, "Inviting organization to collection request");
 
-            const result = await CollectionService.inviteUser({ collectionId: id, organizationId });
-            return createResponseFromResult(c, result);
+            const hasPermission = await ServerPermissionService.isCollectionOwner({
+                user: c.get("user"),
+                collectionId: id,
+            });
+
+            if (!hasPermission) {
+                logger.info({ id }, "User does not have permission to update this collection");
+                return UnauthorizedResponse(c, {
+                    message: "User does not have permission for this action",
+                });
+            }
+
+            const result = await CollectionService.inviteOrganization({
+                collectionId: id,
+                organizationId,
+            });
+
+            return createResponseFromVoidResult(c, result);
         }
     )
     .delete("/:id", permissionMiddleware(Resource.COLLECTION, Action.DELETE), async c => {
         const id = c.req.param("id");
         logger.debug({ id }, "Deleting collection request");
+
+        const hasPermission = await ServerPermissionService.isCollectionOwner({
+            user: c.get("user"),
+            collectionId: id,
+        });
+
+        if (!hasPermission) {
+            logger.info({ id }, "User does not have permission to delete this collection");
+            return UnauthorizedResponse(c, {
+                message: "User does not have permission for this action",
+            });
+        }
 
         const result = await CollectionService.deleteCollection(id);
         return createResponseFromVoidResult(c, result);
